データを検索して、存在した場合のみ削除を実施するように修正

+++ classes/Vizualizer/Plugin/Module/Delete.php
@@ -40,20 +40,23 @@ abstract class Vizualizer_Plugin_Module_Delete extends Vizualizer_Plugin_Module
             $model = $loader->loadModel($name);
             $model->findByPrimaryKey($post[$key]);
 
-            // トランザクションデータベースの取得
-            $connection = Vizualizer_Database_Factory::begin(strtolower($type));
+            // データが検索された場合のみ削除処理を実施
+            if ($model->$key > 0) {
+                // トランザクションデータベースの取得
+                $connection = Vizualizer_Database_Factory::begin(strtolower($type));
 
-            try {
-                $model->delete();
+                try {
+                    $model->delete();
 
-                // エラーが無かった場合、処理をコミットする。
-                Vizualizer_Database_Factory::commit($connection);
+                    // エラーが無かった場合、処理をコミットする。
+                    Vizualizer_Database_Factory::commit($connection);
 
-                $this->removeInput("delete");
-                $this->removeInput($key);
-            } catch (Exception $e) {
-                Vizualizer_Database_Factory::rollback($connection);
-                throw new Vizualizer_Exception_Database($e);
+                    $this->removeInput("delete");
+                    $this->removeInput($key);
+                } catch (Exception $e) {
+                    Vizualizer_Database_Factory::rollback($connection);
+                    throw new Vizualizer_Exception_Database($e);
+                }
             }
         }
     }
